feat(packaging): add portable Windows bundle and development setup
Build x86_64 Windows release archives with checksum-pinned FFmpeg, libplacebo, Vulkan, and build tooling. Generate MSVC import libraries from DLL exports, preserve data imports, and collect runtime dependencies, assets, licenses, and build metadata. Audit PE architecture and transitive dependencies, verify payload hashes, and generate a ZIP with a SHA256 checksum.

Support concurrent positional disk-cache I/O on Windows and accommodate MSVC Vulkan enum types. Resolve portable assets beside the executable, embed the application icon and version resources, and write diagnostics to the user profile for console-free release builds. Preserve existing Linux asset and cache behavior.

Detect Python 3.12 and Visual Studio automatically, add a preparation-only mode, and generate ignored local Cargo configuration for the native SDK. Supply runtime DLL paths through a Cargo runner so cargo run and cargo test work in a fresh terminal without changing the system PATH. Preserve executable arguments and exit codes, and document setup, packaging, and relocation checks.

Cover concurrent cache access and relocated assets, fix Windows newline and accepted-socket behavior in tests, and update fixed-size slice iteration for the current Clippy checks. Validate with 1,919 Rust tests, three Vulkan conversion tests, formatting, Clippy, direct cargo run startup, and an audited release archive containing 16 x86_64 binaries.

+++ src/player/backend/ffmpeg/hw.rs
@@ -263,12 +263,14 @@ fn select_decode_queues(
     })
 }
 
+// bindgen represents these C enums as i32 on MSVC and u32 on GNU targets.
+#[allow(clippy::unnecessary_cast)]
 fn find_queue(
     queue_families: &[vulkan_ffi::AVVulkanDeviceQueueFamily],
     flag: u32,
 ) -> Option<VulkanDecodeQueue> {
     queue_families.iter().find_map(|queue| {
-        let flags = queue.flags;
+        let flags = queue.flags as u32;
         if queue.idx < 0 || queue.num <= 0 || flags & flag == 0 {
             return None;
         }
@@ -304,6 +306,7 @@ pub(super) struct VulkanFrameImages {
     pub(super) planes: Vec<VulkanVideoPlane>,
 }
 
+#[allow(clippy::unnecessary_cast)]
 pub(super) fn vulkan_frame_planes(
     frame: *const ffi::AVFrame,
     raw_format: RawVideoFormat,
@@ -338,8 +341,8 @@ pub(super) fn vulkan_frame_planes(
         }
         planes.push(VulkanVideoPlane {
             image,
-            format: vk_frames.format[plane_index],
-            layout: vk_frame.layout[plane_index],
+            format: vk_frames.format[plane_index] as u32,
+            layout: vk_frame.layout[plane_index] as u32,
             queue_family: vk_frame.queue_family[plane_index],
             semaphore: vk_frame.sem[plane_index] as usize,
             semaphore_value: vk_frame.sem_value[plane_index],
@@ -353,7 +356,7 @@ pub(super) fn vulkan_frame_planes(
         planes.truncate(raw_format.plane_count());
     }
     Ok(VulkanFrameImages {
-        usage: vk_frames.usage,
+        usage: vk_frames.usage as u32,
         planes,
     })
 }
